@@ -1,5 +1,5 @@
 {
-  "name": "gnumoksha/php-freeipa",
+  "name": "pugovok/php-freeipa",
   "type": "library",
   "description": "PHP client to FreeIPA HTTP-RPC",
   "keywords": [
@@ -7,7 +7,7 @@
     "Red Hat Identity Manager",
     "IdM server"
   ],
-  "homepage": "https://github.com/gnumoksha/php-freeipa",
+  "homepage": "https://github.com/pugovok/php-freeipa",
   "license": [
     "LGPL-3.0+"
   ],
@@ -20,35 +20,36 @@
     }
   ],
   "require": {
-    "php": "^7.4",
+    "php": "^8.1",
     "ext-json": "*",
-    "ocramius/package-versions": "*",
+    "laminas/laminas-diactoros": "^2.24",
+    "ocramius/package-versions": "^2.7",
     "php-http/client-common": "^2.0",
     "php-http/discovery": "^1.7",
     "php-http/message": "^1.8",
     "psr/http-client": "^1.0",
     "psr/http-factory": "^1.0",
-    "psr/http-message": "^1.0",
-    "zendframework/zend-diactoros": "^2.1"
+    "psr/http-message": "^1.0"
   },
   "require-dev": {
+    "jetbrains/phpstorm-attributes": "^1.0",
     "php-http/curl-client": "^2.0",
     "php-http/logger-plugin": "^1.1",
     "php-http/mock-client": "^1.3",
-    "phpstan/phpstan": "^0.11.15",
-    "phpstan/phpstan-phpunit": "^0.11.2",
-    "phpunit/phpunit": "^7.5",
-    "squizlabs/php_codesniffer": "*",
-    "vimeo/psalm": "^3.6"
+    "phpstan/phpstan": "^1.9",
+    "phpstan/phpstan-phpunit": "^1.3",
+    "phpunit/phpunit": "^9.5",
+    "squizlabs/php_codesniffer": "^3.7",
+    "vimeo/psalm": "^5.4"
   },
   "autoload": {
     "psr-4": {
-      "Gnumoksha\\FreeIpa\\": "src/"
+      "Pugovok\\FreeIpa\\": "src/"
     }
   },
   "autoload-dev": {
     "psr-4": {
-      "Gnumoksha\\FreeIpa\\": "tests"
+      "Pugovok\\FreeIpa\\": "tests"
     }
   },
   "scripts": {
