@@ -2,11 +2,11 @@
 
 declare(strict_types=1);
 
-namespace Gnumoksha\FreeIpa\Infra\Rpc;
+namespace Pugovok\FreeIpa\Infra\Rpc;
 
-use Gnumoksha\FreeIpa\Infra\Rpc\Response\BodyBuilder as ResponseBodyBuilder;
-use Gnumoksha\FreeIpa\Infra\Rpc\Response\CommonBodyBuilder;
-use Gnumoksha\FreeIpa\Options;
+use Pugovok\FreeIpa\Infra\Rpc\Response\BodyBuilder as ResponseBodyBuilder;
+use Pugovok\FreeIpa\Infra\Rpc\Response\CommonBodyBuilder;
+use Pugovok\FreeIpa\Options;
 use Http\Client\Common\Plugin;
 use Http\Client\Common\Plugin\BaseUriPlugin;
 use Http\Client\Common\Plugin\CookiePlugin;
@@ -23,16 +23,15 @@ use RuntimeException;
  */
 class PluginClientBuilder implements ClientBuilder
 {
-    /** @var \Gnumoksha\FreeIpa\Options */
-    private $options;
-    /** @var \Psr\Http\Client\ClientInterface|null */
-    private $psrHttpClient;
-    /** @var \Psr\Http\Message\UriFactoryInterface */
-    private $uriFactory;
-    /** @var \Gnumoksha\FreeIpa\Infra\Rpc\Response\BodyBuilder */
-    private $responseBodyBuilder;
-    /** @var \Http\Client\Common\Plugin[] */
-    private $httpClientPlugins;
+    private Options $options;
+
+    private ?ClientInterface $psrHttpClient;
+
+    private UriFactoryInterface $uriFactory;
+
+    private ResponseBodyBuilder|CommonBodyBuilder $responseBodyBuilder;
+
+    private array $httpClientPlugins;
 
     public function __construct(
         Options $options,
