@@ -2,16 +2,16 @@
 
 declare(strict_types=1);
 
-namespace Gnumoksha\FreeIpa\Model\Group;
+namespace Pugovok\FreeIpa\Model\Group;
 
-class Group
+final class Group
 {
-    public $cn;
-    public $description;
-    public $gidnumber;
-    public $member;
-    public $memberof;
-    public $memberindirect;
-    public $memberofindirect;
-    public $ipaexternalmember;
+    public string $cn;
+    public string $description;
+    public int $gidnumber;
+    public array $member;
+    public array $memberof;
+    public array $memberindirect;
+    public array $memberofindirect;
+    public array $ipaexternalmember;
 }
