@@ -17,15 +17,16 @@
  * You should have received a copy of the GNU Lesser General Public License
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
+
 declare(strict_types=1);
 
-namespace Gnumoksha\FreeIpa;
+namespace Pugovok\FreeIpa;
 
 use PHPUnit\Framework\TestCase;
 
 class OptionsTest extends TestCase
 {
-    private $options;
+    private Options $options;
 
     public function setUp(): void
     {
