@@ -17,9 +17,10 @@
  * You should have received a copy of the GNU Lesser General Public License
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
+
 declare(strict_types=1);
 
-namespace Gnumoksha\FreeIpa\Infra\Rpc;
+namespace Pugovok\FreeIpa\Infra\Rpc;
 
 /**
  * Represents a class that will build a HTTP RPC Client for FreeIPA.
