@@ -17,11 +17,12 @@
  * You should have received a copy of the GNU Lesser General Public License
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
+
 declare(strict_types=1);
 
-namespace Gnumoksha\FreeIpa\Infra\Rpc\Request;
+namespace Pugovok\FreeIpa\Infra\Rpc\Request;
 
-use Gnumoksha\FreeIpa\Infra\Json\Json;
+use Pugovok\FreeIpa\Infra\Json\Json;
 use PHPUnit\Framework\TestCase;
 
 class CommonBodyTest extends TestCase
